simplified the form type class hierarchy
