+++ Form/ModifySettingsForm.php
@@ -2,6 +2,9 @@
 
 namespace Craue\ConfigBundle\Form;
 
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+
 /**
  * for symfony/form >= 2.8
  *
@@ -9,5 +12,17 @@ namespace Craue\ConfigBundle\Form;
  * @copyright 2011-2015 Christian Raue
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
-class ModifySettingsForm extends AbstractModifySettingsForm {
+class ModifySettingsForm extends AbstractType {
+
+	/**
+	 * {@inheritDoc}
+	 */
+	public function buildForm(FormBuilderInterface $builder, array $options) {
+		$useFqcn = method_exists('Symfony\Component\Form\AbstractType', 'getBlockPrefix');
+
+		$builder->add('settings', $useFqcn ? 'Symfony\Component\Form\Extension\Core\Type\CollectionType' : 'collection', array(
+			'type' => $useFqcn ? 'Craue\ConfigBundle\Form\Type\SettingType' : 'craue_config_setting',
+		));
+	}
+
 }
