+++ Form/Type/SettingType.php
@@ -2,6 +2,10 @@
 
 namespace Craue\ConfigBundle\Form\Type;
 
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
 /**
  * for Symfony >= 2.8
  *
@@ -9,5 +13,28 @@ namespace Craue\ConfigBundle\Form\Type;
  * @copyright 2011-2015 Christian Raue
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
-class SettingType extends AbstractSettingType {
+class SettingType extends AbstractType {
+
+	/**
+	 * {@inheritDoc}
+	 */
+	public function buildForm(FormBuilderInterface $builder, array $options) {
+		$useFqcn = method_exists('Symfony\Component\Form\AbstractType', 'getBlockPrefix');
+
+		$builder->add('name', $useFqcn ? 'Symfony\Component\Form\Extension\Core\Type\HiddenType' : 'hidden');
+		$builder->add('section', $useFqcn ? 'Symfony\Component\Form\Extension\Core\Type\HiddenType' : 'hidden');
+		$builder->add('value', $useFqcn ? 'Symfony\Component\Form\Extension\Core\Type\TextType' : 'text', array(
+			'required' => false,
+		));
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	public function configureOptions(OptionsResolver $resolver) {
+		$resolver->setDefaults(array(
+			'data_class' => 'Craue\ConfigBundle\Entity\Setting',
+		));
+	}
+
 }
